dropitapi/requirements.txt
Django==3.1.5
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
djangorestframework==3.12.2
gunicorn==20.0.4
numpy==1.18.5
Pillow==8.1.0
python-socketio==4.6.1
requests==2.25.1
scikit-learn==0.23.2
scipy==1.5.4
sklearn-crfsuite==0.3.6
websockets==8.1
whitenoise==5.2.0
yfinance==0.1.55
dj_database_url==0.5.0
dnspython==2.1.0
djongo==1.3.6
